+++ src/main/java/arin/ArinException.java
@@ -3,7 +3,7 @@ package arin;
 /**
  * Represents a custom exception for the Arin chatbot.
  */
-class ArinException extends Exception {
+public class ArinException extends Exception {
     /**
      * Constructs an arin.ArinException with the given message.
      *
