+++ src/main/java/arin/command/AddTaskCommand.java
@@ -1,6 +1,10 @@
-package arin;
+package arin.command;
 
+import arin.ArinException;
+import arin.storage.Storage;
 import arin.task.Task;
+import arin.task.TaskList;
+import arin.ui.Ui;
 
 /**
  * Command to add a task (ToDo, Deadline, Event).
