+++ src/main/java/arin/command/Command.java
@@ -1,4 +1,9 @@
-package arin;
+package arin.command;
+
+import arin.ArinException;
+import arin.storage.Storage;
+import arin.task.TaskList;
+import arin.ui.Ui;
 
 /**
  * Represents a command that can be executed.
