+++ src/main/java/arin/task/Deadline.java
@@ -1,8 +1,5 @@
 package arin.task;
 
-import arin.TaskType;
-import arin.task.Task;
-
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
