+++ src/main/java/arin/task/Event.java
@@ -1,8 +1,5 @@
 package arin.task;
 
-import arin.TaskType;
-import arin.task.Task;
-
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
