+++ src/main/java/arin/task/ToDo.java
@@ -1,8 +1,5 @@
 package arin.task;
 
-import arin.TaskType;
-import arin.task.Task;
-
 public class ToDo extends Task {
 
     public ToDo(String description) {
