+++ src/main/java/arin/ui/Ui.java
@@ -1,4 +1,4 @@
-package arin;
+package arin.ui;
 
 import arin.task.Task;
